fix(cron): extract route helpers to sibling lib files (Next.js routing rule)
Vercel build fails on:
  Type error: Route "src/app/api/cron/generate-digest/route.ts" does
  not match the required types of a Next.js Route. "processUser" is
  not a valid Route export field.

Next.js App Router enforces that route.ts files only export HTTP verbs
(GET/POST/etc) plus a small set of config exports — sub-agents added
helpers as `export` for test-import which works against `tsc --noEmit`
but breaks against `next build`. Local tsc didn't catch this; only
the production build does.

Fix:
* New `app/api/cron/generate-digest/process-user.ts` — exports
  processUser, ProcessArgs, ProcessResult, formatPeriodLabel,
  ELIGIBLE_TIERS, BATCH_SIZE, Period
* New `app/api/cron/openrouter-credit-monitor/cycle-metrics.ts` —
  exports computeCycleMetrics, formatCentralTimestamp, CycleMetrics
* Both routes import from the new sibling modules
* Test files re-import from the sibling modules

Also adds 9 env vars to turbo.json's build env passlist
(OPENROUTER_API_KEY, OPENROUTER_LOW_BALANCE_THRESHOLD,
OPENROUTER_ALERT_EMAIL, CRON_SECRET, REDIS_URL, SENTRY_AUTH_TOKEN,
SENTRY_DSN, SENTRY_ORG, SENTRY_PROJECT, STYRBY_SENTRY_MUTED) — Vercel
warned these were set on the project but missing from turbo.json,
which means they don't reach build/cache keys and effectively don't
make it to runtime through Turbo's pipeline.

Local: typecheck clean, 65/65 cron tests pass, `next build` succeeds.
