+++ packages/styrby-web/src/app/api/cron/generate-digest/route.ts
@@ -26,22 +26,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 import crypto from 'crypto';
 import { createAdminClient } from '@/lib/supabase/server';
-import { sendDigestEmail } from '@/lib/resend';
-import { generateDigestContent } from '@/lib/digest/generate';
-
-/**
- * Maximum users processed per cron invocation. Caps Vercel function
- * runtime — at ~1s/user (LLM call dominates) this gives ~5 min headroom.
- */
-const BATCH_SIZE = 200;
-
-type Period = 'daily' | 'weekly';
-
-/** Tiers eligible for each period. Daily = Growth only; Weekly = Pro + Growth. */
-const ELIGIBLE_TIERS: Record<Period, string[]> = {
-  daily: ['growth'],
-  weekly: ['pro', 'growth'],
-};
+import {
+  processUser,
+  ELIGIBLE_TIERS,
+  BATCH_SIZE,
+  type Period,
+} from './process-user';
 
 export async function GET(request: NextRequest) {
   // ---- Auth: timing-safe cron secret comparison ----
@@ -124,105 +114,3 @@ export async function GET(request: NextRequest) {
   return NextResponse.json({ generated, emailed, errors });
 }
 
-interface ProcessArgs {
-  supabase: ReturnType<typeof createAdminClient>;
-  userId: string;
-  period: Period;
-  periodStart: Date;
-  periodEnd: Date;
-}
-
-interface ProcessResult {
-  generated: boolean;
-  emailed: boolean;
-  reason?: string;
-}
-
-/**
- * Generate + persist + email a digest for a single user.
- * Returns granular flags so the orchestrator can tally counts.
- */
-export async function processUser(args: ProcessArgs): Promise<ProcessResult> {
-  const { supabase, userId, period, periodStart, periodEnd } = args;
-
-  // Pull this user's sessions in the window.
-  const { data: sessions, error: sessionsError } = await supabase
-    .from('sessions')
-    .select('id, title, agent_type, status, total_cost_usd, message_count, created_at')
-    .eq('user_id', userId)
-    .gte('created_at', periodStart.toISOString())
-    .lte('created_at', periodEnd.toISOString())
-    .order('created_at', { ascending: false })
-    .limit(200);
-
-  if (sessionsError) {
-    return { generated: false, emailed: false, reason: sessionsError.message };
-  }
-
-  // Skip silent: no sessions = no spam.
-  if (!sessions || sessions.length === 0) {
-    return { generated: false, emailed: false, reason: 'no_sessions' };
-  }
-
-  // ---- LLM digest generation ----
-  const content = await generateDigestContent({ period, sessions });
-
-  // ---- Persist (idempotent on UNIQUE (user_id, period, period_start)) ----
-  // ON CONFLICT DO NOTHING via upsert with ignoreDuplicates.
-  const { error: insertError } = await supabase.from('digest_summaries').upsert(
-    {
-      user_id: userId,
-      period,
-      period_start: periodStart.toISOString(),
-      period_end: periodEnd.toISOString(),
-      session_count: sessions.length,
-      content,
-    },
-    { onConflict: 'user_id,period,period_start', ignoreDuplicates: true }
-  );
-
-  if (insertError) {
-    return { generated: false, emailed: false, reason: insertError.message };
-  }
-
-  // ---- Email dispatch ----
-  // Look up email — we use the admin auth API to keep the schema lean
-  // (no need to denormalise email into subscriptions/profiles).
-  const { data: userResp } = await supabase.auth.admin.getUserById(userId);
-  const email = userResp?.user?.email;
-  if (!email) {
-    return { generated: true, emailed: false, reason: 'no_email' };
-  }
-
-  let emailed = false;
-  try {
-    const dateLabel = formatPeriodLabel(period, periodStart, periodEnd);
-    await sendDigestEmail({
-      email,
-      period,
-      dateLabel,
-      sessionCount: sessions.length,
-      content: content ?? '',
-    });
-    emailed = true;
-    await supabase
-      .from('digest_summaries')
-      .update({ emailed_at: new Date().toISOString() })
-      .eq('user_id', userId)
-      .eq('period', period)
-      .eq('period_start', periodStart.toISOString());
-  } catch (err) {
-    return { generated: true, emailed: false, reason: (err as Error).message };
-  }
-
-  return { generated: true, emailed };
-}
-
-/**
- * Format a human label for the digest period (used in the email subject
- * and dashboard panel). Example: "week of May 4" or "May 4".
- */
-function formatPeriodLabel(period: Period, start: Date, end: Date): string {
-  const monthDay = end.toLocaleDateString('en-US', { month: 'long', day: 'numeric' });
-  return period === 'weekly' ? `week of ${monthDay}` : monthDay;
-}
